Work towards history, fix render bugs
If the LinearProgressIndicator comes first in the stack, this
causes a number of exceptions to be thrown. This seems
related to the SearchResults widget changing position in the
stack as the progress bar appears and disappears. Making
the SearchResult widget the first widget resolves it. This is
also correct because we want the progress bar to be overlaid
OVER the results, and Stack renders widgets from bottom to
top in the order provided.